Remove helm-project-operator lock
The lock on the helm-project-operator doesn't make sense since the operator may not be watching for ProjectHelmCharts in the system namespace if ProjectLabel and/or SystemProjectLabelValue are provided. Therefore, locking on helm-controller and helm-locker is sufficient since it does watch for HelmCharts and HelmReleases only in the systemNamespace, which means that multiple Project Operators will still not be deployable in the same system namespace as the Helm Project Operator.